Add x-clacks-overhead
This adds a non-standard header, that was inspired by the Discworld series. There a series of towers are used to constantly re-transmit the memory of a lost son, throughtout a network. Inspired by this many servers serve a similar header with memories of lost comrades (including sir terry pratchett, the author).

Here I propose adding Simon and James to the clacks-overhead served by EU, a small way to remember them.

> In Sir Terry's novel "Going Postal", the story explains that the
> inventor of the Clacks - a man named Robert Dearheart, lost his only son
> in a suspicious workplace accident, and in order to keep the memory of
> his son alive, he transmitted his son's name as a special operational
> signal through the Clacks to forever preserve his memory:

https://xclacksoverhead.org/home/about