Update bashrc role
Updates:
1. This update will now allow us to add the bashrc content to more than 1
user. However, the default behaviour is retained (for backwards
compatibility) i.e., by default it adds the contents to the galaxy user.
2. Added ENVs required for certain gxadmin commands
3. Added new change_to_jwd function that uses the new get_jwd bash
script. This script and the function accepts only the job id. The new
alias is chg2jwd and also the old method chg2wd is available just in
case.