Switch Keras to TPV
Type of compute resource (CPU or GPU) for the keras_train_eval tool depends on the user's selection. The processing of the user's selection has been moved from sorting hat to TPV.

The difference with respect to commit c337c40 (the previous attempt) is that nothing in files/galaxy/dynamic_rules/usegalaxy/sorting_hat.py is deleted this time.