Migrate HTCondor manager from sn05 to sn06
Since sn05 will be upgraded to Rocky 9.0 we cannot use old version of
HTCondor anymore. So, we migrate the central manager from sn05 to sn06
until the cloud is also updated to Rocky 9.0. after which we can roll
out HTCondor 10 and setup HTCondor manger on sn05 and retire it on sn06.