HTCondor clusters playbook
This commit adds a playbook (htcondor.yml) that manages both the primary and secondary HTCondor clusters. The idea is to take advantage of Ansible's group and host vars features to configure both clusters with little verbosity and great flexibility. It is also meant to replace the way HTCondor is being configured at the moment (the sn06.yml playbook).

See https://github.com/usegalaxy-eu/infrastructure-playbook/pull/951 for an in-depth explanation of the motivation behind the changes.