Enable TCP forwarding on the headnode exclusively for 127.0.0.1:2222
The sshd that the systemd-nspawn container for the HTCondor migration runs listens on 127.0.0.1:2222 (on the headnode).

Ansible's ProxyCommand setting, specified in the Ansible inventory file (hosts), requires TCP forwarding to work.

This PR enables TCP forwarding, but only local forwarding and exclusively to the container's sshd port.

In addition, the changes are tied to the `central-manager-secondary-host.yml` group vars file and to the `central-manager-secondary-host` group from the Ansible inventory, so that when the migration is complete and the container is removed, the setting no longer applies.

For more information, see https://manpages.debian.org/bookworm/openssh-server/sshd_config.5.en.html#PermitOpen and https://manpages.debian.org/bookworm/openssh-server/sshd_config.5.en.html#AllowTcpForwarding.