Configure "galaxy" user and group in the HTCondor cluster
Please remember that at the moment, this commit only affects the secondary HTCondor cluster, because the htcondor.yml playbook excludes `sn06.galaxyproject.eu` (which plays the central manager and submit roles in the primary cluster), and the worker nodes are not managed via this repository.

Run `usegalaxy_eu.handy.os_setup` on all members of the HTCondor cluster. Configure the group variables so that the role frees up uid 999 and gid 999 and creates a "galaxy" user with those identifiers.

Additionally, `galaxy_user` and `galaxy_group` are now defined globally (for all hosts).