Change definition of "stuck VM" in manage_vms script
In the current version of the script, stuck VMs are those whose name starts with `vgcnbwc-worker-` and do not show up in `condor_status`.

Change the definition of stuck so that stuck VMs are those that do not reply to ICMP echo requests.

This prevents the script from removing machines belonging to the secondary HTCondor cluster.