Patch galaxy_jwd.py to return the correct JWD for embedded Pulsar jobs running on the secondary cluster
The embedded Pulsar job runner for the secondary cluster is called `pulsar_embedded_secondary`. Since galaxy_jwd.py uses different logic for getting the JWD of embedded pulsar jobs but only applies it when `job_runner_name == "pulsar_embedded"`, it yields the wrong job working directories for jobs whose runner is `pulsar_embedded_secondary`.

This patch works around the issue so that the stop-ITs script works on the secondary cluster. It does NOT fix it properly, because if all embedded pulsar runners do not use the same staging dir then it breaks.