Invert query order when removing orphan Condor jobs
Originally, the script first gets the jobs from the Galaxy database, then from Condor. If a new job is submitted in between the two queries, it will get killed.

Arguably this is very unlikely, but over the years such sources of mysterious failure accumulate and things get messy.

Invert the order of the queries so that the problem cannot occur in the first place.