Query Galaxy database also for queued jobs when removing orphan Condor jobs
While testing the script in dry-run mode, I have spotted that in some cases it removes jobs that `gxadmin report job-info` reports as being in "running" state.

When a Galaxy job goes into queued state, it is sent to HTCondor via `condor_submit`, entering the HTCondor queue in "idle" state. When the job enters the running state in HTCondor, then its state also changes to "running" in Galaxy.

 However, the state change in Galaxy is not instantaneous. This means that a job can be running in HTCondor, yet Galaxy does not necessarily have to know about it (it would still be shown as "queued" in Galaxy). This is precisely the case that results in the removal of a job that is in fact running.