Add comment mentioning the idea to manage dependencies of roles differently
It is a time sink and very annoying to have to re-run playbooks until no more "command not found" errors appear because of forgetting to install role dependencies (meaning system packages, not other roles). See issue https://github.com/galaxyproject/ansible-gxadmin/issues/3 for an example of someone complaining about this recurring problem.

I thought of creating a role called "usegalaxy_eu.packages" that alleviates this problem. The idea is to define a list of the packages each role depends on in group_vars/all.yml (see the example below), then have the role check what other roles are imported in a play and install the corresponding dependencies.

```yaml
# Ansible role package dependencies (managed by usegalaxy_eu.packages)
packages:
  usegalaxy_eu.handy.os_setup:
    - findutils
  galaxyproject.gxadmin:
    - git
    - make
    - "postgresql{{ '-client' if ansible_os_family == 'Debian' }}"
  usegalaxy-eu.bashrc:
    - python3-psycopg2
    - python3-pyyaml
```

The idea to create this role stems just from stumbling upon the problem. See https://github.com/galaxyproject/ansible-gxadmin/pull/11 for an additional discussion on the topic.