Use role `grycap.htcondor` from usegalaxy-eu organization
As part of the migration to HTCondor 23, the role managing HTCondor was switched from `usegalaxy_eu.htcondor` to `grycap.htcondor`. I forked it and added linting and a couple of tests on top, and that fork is what useGalaxy.eu is running today.

However, the entry for `grycap.htcondor` in requirements.yaml still points to that fork https://github.com/kysrpex/grycap-ansible-role-htcondor on my personal account. This is wrong for many reasons.

This commit changes the source URL of the role to a fork on the usegalaxy-eu organization.