Workaround issue with RSpace file source template
Galaxy does not install the `rspace-client` dependency for the RSpace file source because it ignores file source templates when determining which conditional dependencies should be installed. See Galaxy issue https://github.com/galaxyproject/galaxy/issues/20483.

Install it to the Galaxy venv via the `hxr.install-to-venv` role as a workaround. Write comment referencing the issue and stating that the change can be reverted when the issue is fixed.