Update Seurat preprocessing memory
There was a error report on EU server from a user trying to run this tool on a dataset of over 600MB. We cannot yet estimate the memory based on the input file size. I want to test whether it works out for now. Later, we can implement a proper rule on shared TPV.