Update authorized persons and encrypt their public keys
Adjust the authorized persons configuration based on an exhaustive examination of the contents of the authorized keys files for each server. It does not match the current status of each server (that would be a mistake), but provides a reasonably similar (and updated access policy) that can be tweaked within the coming months.

In addition, use `key` instead of `keys` (the latter is a reserved Ansible keyword) and encrypt the public keys of all users.