Trust sn12 SSH host key
Node sn12 cannot be managed via Jenkins because its host key is not trusted (it is not signed by the usegalaxy.eu cert authority).

Although the solution is to sign it, as a workaround add it to the `known_hosts` Makefile target so that PR #1673 (Centralized SSH key management, https://github.com/usegalaxy-eu/infrastructure-playbook/pull/1673) can be merged.