Switch rustus directory to `misc06` on zfs06 box
Change the rustus uploads directory to `misc06` on the zfs07 box with the goal of rebooting the zfs07 box to deal with stuck delegations.

This doesn't completely prevent uploads downtime, but it makes it "instantaneous"; only ongoing uploads break. The other solution (just rebooting) arguably allows newer uploads to continue after the reboot. It's a tradeoff between small-file uploads and large file uploads, and all options are bad. Choose your poison.

A real solution that prevents downtime is to store nfo files in a redis instance setup for high-availability and S3-like storage (e.g. NetApp).