Exclude `tmpwatch` scripts from `usegalaxy-eu.remove-orphan-condor-jobs` role
Add a configuration variable `remove_orphan_jobs_exclude` to `usegalaxy-eu.remove-orphan-condor-jobs` to exclude jobs containing specific strings from the cleanup.

Define this variable in the group vars for the `maintenance` in a way such that it matches the basenames of the scripts that clean up old files from the job working directories.