Consider only jobs launched by Galaxy as candidates for orphan job removal
The role `usegalaxy-eu.remove-orphan-condor-jobs` finds jobs that are known to HTCondor but not to Galaxy anymore, e.g. because the jobs were removed with `gxadmin mutate fail-job`, then kills the jobs.

Specifically, it considers any HTCondor jobs that do not show up in the Galaxy database as orphans kills them. This interferes with HTCondor jobs that have not been launched by Galaxy (e.g. filesystem maintenance jobs).

Filter the list of HTCondor jobs using a constraint so that only jobs launched by Galaxy can be flagged as orphan and cancelled.

Co-authored-by: Mira <86979912+mira-miracoli@users.noreply.github.com>